refactor: 💡 rename settings to data folder
