+++ app/src/main/java/com/rickyhu/hushkeyboard/data/AppSettings.kt
@@ -1,7 +1,5 @@
-package com.rickyhu.hushkeyboard.settings
+package com.rickyhu.hushkeyboard.data
 
-import com.rickyhu.hushkeyboard.settings.options.ThemeOption
-import com.rickyhu.hushkeyboard.settings.options.WideNotationOption
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -11,3 +9,14 @@ data class AppSettings(
     val vibrateOnTap: Boolean = true,
     val wideNotationOption: WideNotationOption = WideNotationOption.WideWithW
 )
+
+enum class ThemeOption { System, Light, Dark }
+
+enum class WideNotationOption {
+    WideWithW, WideWithLowercase;
+
+    override fun toString() = when (this) {
+        WideWithW -> "Use w (Rw)"
+        WideWithLowercase -> "Use lowercase (r)"
+    }
+}
