--- app/src/main/java/com/rickyhu/hushkeyboard/settings/options/ThemeOption.kt
@@ -1,3 +1,0 @@
-package com.rickyhu.hushkeyboard.settings.options
-
-enum class ThemeOption { System, Light, Dark }
