--- app/src/main/java/com/rickyhu/hushkeyboard/settings/options/WideNotationOption.kt
@@ -1,10 +1,0 @@
-package com.rickyhu.hushkeyboard.settings.options
-
-enum class WideNotationOption {
-    WideWithW, WideWithLowercase;
-
-    override fun toString() = when (this) {
-        WideWithW -> "Use w (Rw)"
-        WideWithLowercase -> "Use lowercase (r)"
-    }
-}
