+++ app/src/main/java/com/rickyhu/hushkeyboard/ui/keyboard/NotationKeyButtonRow.kt
@@ -8,8 +8,8 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.rickyhu.hushkeyboard.data.options.WideNotationOption
 import com.rickyhu.hushkeyboard.model.CubeKey
-import com.rickyhu.hushkeyboard.settings.options.WideNotationOption
 import com.rickyhu.hushkeyboard.ui.keyboard.buttons.NotationKeyButton
 
 @Composable
